### 2021-10-15 #474 @quishqa
qualR:  An R package to download Sao Paulo and Rio de Janeiro air pollution data
Submitting Author Name: Mario Gavidia-Calderón
Submitting Author Github Handle: @quishqa
Other Package Authors Github handles:  @Schuch666, @mftandra
Repository:  https://github.com/quishqa/qualR
Version submitted:0.9.5
Submission type: Standard

Version accepted: TBD

---

```
Package: qualR
Title: An R package to download Sao Paulo and Rio de Janeiro air pollution data
Version: 0.9.5
    person(given = "Mario",
           family = "Gavidia-Calderón",
           email = "mario.calderon@iag.usp.br",
    person(given = "Maria de Fatima",
           family = "Andrade",
           role = c("ctb", "ths"),
           email = "maria.andrade@iag.usp.br",
    person(given = "Daniel",
           family = "Schuch",
           role = c("aut","ctb"),
           email = "d.schuch@northeastern.edu",
           comment = c(ORCID = "0000-0001-5977-4519")))
Description: A package to download information from CETESB QUALAR and
    MonitorAr systems. It contains function to download different parameters, a set of
    criteria pollutants and the most frequent meteorological parameters used in
    air quality data analysis and air quality model evaluation.
    XML,
    httr,
    jsonlite
    covr,
```

Sao Paulo data is available through [CETESB QUALAR System](https://qualar.cetesb.sp.gov.br/qualar/home.do) which limits the download to one parameter for one air quality station for one year, Rio de Janeiro is available through [data.rio.API](https://www.data.rio/datasets/dados-hor%C3%A1rios-do-monitoramento-da-qualidade-do-ar-monitorar/explore) which is not so user-friendly. `qualR` downloads multiple parameters for different air quality stations and produces completed ready-to-use data frames (missing hours are padded out with NA) with a date column in POSIXct type that allows temporal aggregation and compatibility with `openair` package.

Any researchers that work with air quality and weather data of Sao Paulo and Rio de Janeiro: Meteorologists, Epidemiologists, Environmental engineers, postgraduates students that work in air quality modeling and field measurements campaigns, also stakeholders and the community. This packages facilitates the retrival of Sao Paulo and Rio de Janeiro air quality data, facilitates exploratory data analysis, and enhance code reproducibility.

[`koffing`](https://github.com/atmoschem/koffing) retrieves Sao Paulo State air quality data.  `qualR` offers more functionalities (i.e. allows retrieve many parameters from one air quality station, accepts pollutants abbreviations and air quality station names in functions parameters, etc), and returns complete datasets ready to use (no missing hours, concentration in numeric format). `qualR` also include Rio de Janeiro city and the location of each air quality stations.  `qualR` is more user friendly and it is actively maintainded.
