+++ CONTRIBUTING.md
@@ -1,0 +1,12 @@
+# How to contribute
+We welcome any suggestion, requests and bug report through the [issue tracker](https://github.com/quishqa/qualR/issues).
+
+## Any bugs?
+- If you found any issue while using `qualR`, submit an issue [here](https://github.com/quishqa/qualR/issues).
+- Remember to add a reproducible example and the output of `sessionInfo()`.
+
+## A little bit shy?
+- You can email the person as maintainer in the `DESCRIPTION` file (in portuguese or english).
+
+
+Thanks for your contribution :)
