+++ DESCRIPTION
@@ -25,7 +25,7 @@ License: MIT + file LICENSE
 Encoding: UTF-8
 LazyData: true
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.1.1
+RoxygenNote: 7.1.2
 Imports: 
     XML,
     httr,
@@ -33,8 +33,11 @@ Imports:
 URL: https://github.com/quishqa/qualR
 BugReports: https://github.com/quishqa/qualR/issues
 Suggests:
+    knitr,
     covr,
-    testthat (>= 3.0.0)
+    testthat (>= 3.0.0),
+    rmarkdown
 Depends: 
     R (>= 3.5.0)
 Config/testthat/edition: 3
+VignetteBuilder: knitr
