+++ R/data.R
@@ -1,39 +1,69 @@
 #' CETESB AQS station latitude and longitude
 #'
-#' List of CETESB QUALAR AQS latitudes and longitudes. Use this to check the
-#' AQS location, if you need to make some maps. AQS names contains diacritics.
+#' List of CETESB QUALAR air qualitty stations (AQS) latitudes and longitudes.
+#' Use this to check the AQS \code{aqs_code} argument in CetesbRetrieveParam()
+#' function. AQS names are without diacritics.
 #'
-#'
-#' @format A data frame with 4 variables: \code{name} , \code{code}, \code{lat}, and
-#' \code{latitude} of AQS.
+#' @format A data frame with 68 observations and 4 variables:
+#' \describe{
+#' \item{name}{CETESB AQS name.}
+#' \item{code}{CETESB AQS code in QUALAR System.}
+#' \item{lat}{CETESB AQS latitude.}
+#' \item{lon}{CETESB AQS lontgitude.}
+#' }
+#' @examples
+#' cetesb_aqs
 "cetesb_aqs"
 
 #' CETESB Parameters
 #'
 #' List of CETESB QUALAR available parameters and units. Use this to check the
-#' pol_code parameter.
+#' \code{parameters} argument.
 #' Parameter names are without diacritics.
 #'
-#' @format A data frame with 3 variables: \code{name}, \code{units}, and \code{code}
-#'   of parameter.
+#' @format A data frame with 20 observations and 3 variables:
+#' \describe{
+#' \item{name}{CETESB QUALAR paramater abbreviation and name.}
+#' \item{units}{Parameter units.}
+#' \item{code}{Parameter CETESB QUALAR code.}
+#' }
+#' @examples
+#' cetesb_param
 "cetesb_param"
 
 #' Monitor Ar AQS stations.
 #'
 #' List of Monitor Ar Rio air quality stations (AQS) with their codes
-#' and locations. Use this to check the aqs_code parameter in
-#' MonitorArRetrieve() function.
+#' and locations. Use this to check the parameters argument in
+#' MonitorArRetrieveParam() function.
 #'
-#' @format A data frame with 6 variables: \code{name}, \code{code}, \code{lon}, \code{lat},
-#' \code{X_UTM_Sirgas2000}, and  \code{Y_UTM_Sirgas2000}
-#'   of AQS.
+#' @format A data frame with 8 observation and 6 variables:
+#' \describe{
+#' \item{name}{MonitorAr Program AQS name.}
+#' \item{code}{MonitorAr Program AQS abbreviation.}
+#' \item{lon}{MonitorAr Program AQS longitude.}
+#' \item{lat}{MonitorAr Program AQS latitude.}
+#' \item{X_UTM_Sirgas2000}{MonitorAr Program AQS longitude in SIRGAS 2000
+#'  (EPSG:31983).}
+#' \item{Y_UTM_Sirgas2000}{MonitorAr Program AQS latitude in SIRGAS 2000
+#'  (EPSG:31983).}
+#' }
+#' @examples
+#' monitor_ar_aqs
 "monitor_ar_aqs"
 
 #' Monitor Ar Parameters
 #'
 #' List of Monitor Ar Rio available parameters. Use this to check the
-#' param parameter of MonitorArRetrieve() function .
+#' parameters argument in MonitorArRetrieveParam() function.
+#' Parameter names are without diacritics.
 #'
-#' @format A data frame with 3 variables: \code{code}, \code{name}, and \code{units}
-#'   of parameter.
+#' @format A data frame with 18 observations and 3 variables:
+#' \describe{
+#' \item{code}{MonitorAr parameter abbreviation or code.}
[…]
