### 2021-10-19 4b8f136
including on.exit in tests of testthat
