+++ tests/testthat/test-CetesbRetrieveParam.R
@@ -1,5 +1,8 @@
 test_that("CetesbRetrieveParam works!", {
 
+  current_folder <- getwd()
+  on.exit(setwd(current), add = T)
+
   load("sysdata.rda")
 
   dir.create(file.path(tempdir(), "cetesb-data2"))
@@ -8,8 +11,6 @@ test_that("CetesbRetrieveParam works!", {
   pin_param <- CetesbRetrieveParam(u,p,c("o3","NOX", "VV"), 99,
                                    "01/01/2020", "07/01/2020",
                                    to_csv = TRUE)
-
-
 
   expect_equal(2 * 2, 4)
 })
