+++ tests/testthat/test-MonitorArRetrieve.R
@@ -1,7 +1,7 @@
 test_that("MonitorArRetrieve works", {
 
   current_folder <- getwd()
-  on.exit(setwd(current), add = T)
+  on.exit(setwd(current_folder), add = T)
 
   dir.create(file.path(tempdir(), "cetesb-data"))
   setwd(file.path(tempdir(), "cetesb-data"))
