+++ R/WriteCSV.R
@@ -14,7 +14,7 @@ WriteCSV <- function(aqs_data_df, aqs_name, start_date, end_date, vars){
                       paste0(vars, collapse = "_"), "_",
                       gsub("/", "-", start_date), "_",
                       gsub("/", "-", end_date), ".csv")
-  utils::write.table(aqs_data_df, file_name, sep = ",", row.names = F )
+  utils::write.table(aqs_data_df, file_name, sep = ",", row.names = FALSE)
 
   file_path <- paste(getwd(), file_name, sep = "/")
   cat(paste(file_path, "was created \n"))
