+++ R/CetesbRetrieveMetPol.R
@@ -112,7 +112,7 @@ CetesbRetrieveMetPol <- function(username, password,
                          pm10 = pm10$pol,
                          pm25 = pm25$pol,
                          aqs = o3$aqs,
-                         stringsAsFactors = F)
+                         stringsAsFactors = FALSE)
   cat(paste(
     "Download complete for", unique(all_data$aqs), "\n"
   ))
