### 2021-10-19 cb28390
Fix long code lines in CetesbRetrieve.R
