+++ R/MonitorArRetrieveParam.R
@@ -1,12 +1,13 @@
 #' Download air quality and meteorology information from MonitorAr-Rio
 #'
-#' This function download air quality and meteorology measurements from MonitorAr-Rio
-#' program from Rio de Janeiro city.
+#' This function download air quality and meteorology measurements from
+#' MonitorAr-Rio program from Rio de Janeiro city.
 #'
 #' @param start_date Date to start downloading in dd/mm/yyyy
 #' @param end_date Date to end downloading in dd/mm/yyyy
 #' @param aqs_code AQS code
-#' @param parameters Paremeters to download. It can be a vector with many parameters.
+#' @param parameters Paremeters to download.
+#' It can be a vector with many parameters.
 #' @param to_local Date information in local time. TRUE by default.
 #' @param verbose Print query summary.
 #' @param to_csv Creates a csv file. FALSE by default
@@ -31,12 +32,12 @@ MonitorArRetrieveParam <- function(start_date, end_date, aqs_code, parameters,
   # Check if params are measured
   if (!prod(parameters %in% param_monitor_ar$code)){
     stop("One or all wrong param codes, please check monitor_ar_param", # nocov
-         call. = FALSE)                                                           # nocov
+         call. = FALSE)                                                 # nocov
   }
 
   if (!(aqs_code %in% aqs_monitor_ar$code )){
-    stop("Wrong aqs_code, please check monitor_ar_aqs",                           # nocov
-         call. = FALSE)                                                           # nocov
+    stop("Wrong aqs_code, please check monitor_ar_aqs",                 # nocov
+         call. = FALSE)                                                 # nocov
   }
 
   aqs_name <- aqs_monitor_ar$name[aqs_monitor_ar$code == aqs_code]
@@ -49,8 +50,10 @@ MonitorArRetrieveParam <- function(start_date, end_date, aqs_code, parameters,
     cat("Period: From", start_date, "to", end_date, "\n")
   }
 
-  start_date_format <- as.POSIXct(strptime(start_date, format="%d/%m/%Y"), tz = "UTC")
-  end_date_format <-  as.POSIXct(strptime(end_date, format="%d/%m/%Y"), tz = "UTC")
+  start_date_format <- as.POSIXct(strptime(start_date, format="%d/%m/%Y"),
+                                  tz = "UTC")
+  end_date_format <-  as.POSIXct(strptime(end_date, format="%d/%m/%Y"),
+                                 tz = "UTC")
 
   # Function to create the WHERE query
   where_query <- function(ds, de, aqs){
@@ -68,14 +71,18 @@ MonitorArRetrieveParam <- function(start_date, end_date, aqs_code, parameters,
   if (length(parameters) == 1){
     outfields <- paste("Data", parameters, sep = ",")
   } else {
-    outfields <- paste("Data", paste(parameters, collapse = ","), sep = ",")  # nocov
+    outfields <- paste("Data", paste(parameters, collapse = ","),
+                       sep = ",")  # nocov
   }
 
-  url <- "https://services1.arcgis.com/OlP4dGNtIcnD3RYf/arcgis/rest/services/Qualidade_do_ar_dados_horarios_2011_2018/FeatureServer/2/query?"
+  url <- paste0("https://services1.arcgis.com/OlP4dGNtIcnD3RYf/arcgis/rest/",
+                "services/Qualidade_do_ar_dados_horarios_2011_2018/",
+                "FeatureServer/2/query?")
 
   res <- httr::GET(url,
                    query = list(
-                     where = where_query(start_date_format, end_date_format, aqs_code),
+                     where = where_query(start_date_format, end_date_format,
+                                         aqs_code),
[…]
