### 2021-10-20 0c68af2
add xlim and ylim to plot in vignette to alow check the package without the data
