### 2021-11-29 #474 @beatrizmilz
Hi! Thank you for the invitation to review the package @ldecicco-USGS and congratulations to the authors!

Here is my review:

Estimated hours spent reviewing: 6h

---

I wrote more extensive comments in the [full report](https://github.com/beatrizmilz/ropensci_reviews/blob/main/review-qualR/README.md), made with `ropensci-org/pkgreviewr`.

Here are some key points that would be interesting to work on, to improve the package:

- Consistent naming styles: it would be better to choose a language (English), and a coding style (ex. snake case) and use them in all function names, arguments, and names of columns in the datasets returned.

- The tests can be improved. Consider also using GitHub Actions to perform an automated check.

- Files in the root directory: some of the files do not seem fit to be stored there (such as the files related to the paper), and some need to be ignored in the build of the package. I wrote in more detail in the report.

- The spell-check found some typos that could be corrected.

- I wrote in the report some ideas for the vignettes, but they are only suggestions to make it easier for people to use the data.

- Modifying directories using `setwd()` is not something recommended in the book R Packages ( https://r-pkgs.org/r.html ).

Report: https://github.com/beatrizmilz/ropensci_reviews/blob/main/review-qualR/README.md
