### 2021-12-01 #474 @quishqa
Hi @beatrizmilz thanks a lot for your detailed review, we'll be solving all the observations in the following days 👍🏽
