### 2021-12-14 #474 @kauedesousa
Dear @ldecicco-USGS and @quishqa here goes my comments on the package. This is a nice one. Congrats!!

---

* Datasets `monitor_ar_aqs` and `monitor_ar_param` don't match with the general naming used for the functions. I suggest that you rename these files to match with your naming logic. You may rename like, `ArStation` and `ArParameters`
* Same for the CETESB data sets
* Part of your functions (for CETESB) need a previous account with CETESB, and you only mention that by the end of your vignette. You need to mention that on the description of the functions `CetesbRetrieveMet()` and the other related.
* Your vignette don't run smoothly because you call the package `openair` with `::` without explicitly mentioning it with `library(openair)`. You should call the package with `library()`.
* In line 114 of `/R/CetesbRetrieve.R` you start with an `if()` and then a `else` with another `if()`. I think that these statements are very difficult to understand and to fix. Maybe you can use only `if()`s and stress on the possible outcomes that the call may return to you.
* You test if the functions are returning the right dimensions of your expected data frame. But you don't test whether the outcome is the correct data. Maybe you can download a small data set from the server and put it in the `sysdata.rda` and validate it with the output of your functions.
* Fix the typo in line 16 of file `/R/CetesbRetrieveMet.R`
* I recommend to put the paper files in /inst/paper so your main directory is more organized. Same for the png files.
* Remove `README.Rmd` and keep only `README.md`
* /data-raw/ could be placed inside /data/
* Remove LICENSE.md
* In the DESCRIPTION file, put the links for ETESB QUALAR and MonitorAr systems if any inside a <>, like this <https://cetesb.sp.gov.br/ar/qualar/> is very likely that CRAN will ask you to do that.
* Make a file called qualR.R in the \R directory, it makes a nice description of your package. There are some complex versions like this <https://github.com/ropensci/rgbif/blob/master/R/rgbif-package.r> or simplified versions like this <https://github.com/ropensci/chirps/blob/master/R/chirps.R>
* The package is nice and delivers what is promissed. Congrats!!

### Session info

`````
─ Session info  🍰  👨‍⚕️  📱   ────────────────────────────────────────────────────────────────────────────────────────
 hash: shortcake, man health worker, mobile phone

 os       macOS 12.0.1
 system   x86_64, darwin17.0
 ui       RStudio
 collate  en_US.UTF-8
 ctype    en_US.UTF-8
 tz       Europe/Oslo
 rstudio  1.4.1716 Juliet Rose (desktop)
 pandoc   2.9.1.1 @ /usr/local/bin/pandoc

─ Packages ──────────────────────────────────────────────────────────────────────────────────────────────────────────
 package     * version  date (UTC) lib source
 cachem        1.0.6    2021-08-19 [1] CRAN (R 4.1.0)
 callr         3.7.0    2021-04-20 [1] CRAN (R 4.1.0)
 cli           3.1.0    2021-10-27 [1] CRAN (R 4.1.0)
 crayon        1.4.2    2021-10-29 [1] CRAN (R 4.1.0)
 curl          4.3.2    2021-06-23 [1] CRAN (R 4.1.0)
 desc          1.4.0    2021-09-28 [1] CRAN (R 4.1.0)
 devtools    * 2.4.2    2021-06-07 [1] CRAN (R 4.1.0)
 ellipsis      0.3.2    2021-04-29 [1] CRAN (R 4.1.0)
 fastmap       1.1.0    2021-01-25 [1] CRAN (R 4.1.0)
 fs            1.5.0    2020-07-31 [1] CRAN (R 4.1.0)
 glue          1.4.2    2020-08-27 [1] CRAN (R 4.1.0)
 httr          1.4.2    2020-07-20 [1] CRAN (R 4.1.0)
 jsonlite      1.7.2    2020-12-09 [1] CRAN (R 4.1.0)
 knitr         1.36     2021-09-29 [1] CRAN (R 4.1.0)
 lifecycle     1.0.1    2021-09-24 [1] CRAN (R 4.1.0)
 magrittr      2.0.1    2020-11-17 [1] CRAN (R 4.1.0)
 memoise       2.0.0    2021-01-26 [1] CRAN (R 4.1.0)
 pkgbuild      1.2.0    2020-12-15 [1] CRAN (R 4.1.0)
 pkgload       1.2.3    2021-10-13 [1] CRAN (R 4.1.0)
 prettyunits   1.1.1    2020-01-24 [1] CRAN (R 4.1.0)
 processx      3.5.2    2021-04-30 [1] CRAN (R 4.1.0)
 ps            1.6.0    2021-02-28 [1] CRAN (R 4.1.0)
 purrr         0.3.4    2020-04-17 [1] CRAN (R 4.1.0)
 qualR       * 0.9.5    2021-12-14 [1] Github (quishqa/qualR@dc15bd0)
 R6            2.5.1    2021-08-19 [1] CRAN (R 4.1.0)
 remotes       2.4.1    2021-09-29 [1] CRAN (R 4.1.0)
 rlang         0.4.12   2021-10-18 [1] CRAN (R 4.1.0)
 roxygen2    * 7.1.2    2021-09-08 [1] CRAN (R 4.1.0)
 rprojroot     2.0.2    2020-11-15 [1] CRAN (R 4.1.0)
 rstudioapi    0.13     2020-11-12 [1] CRAN (R 4.1.0)
 sessioninfo   1.2.1    2021-11-02 [1] CRAN (R 4.1.0)
 stringi       1.7.5    2021-10-04 [1] CRAN (R 4.1.0)
 stringr       1.4.0    2019-02-10 [1] CRAN (R 4.1.0)
 testthat    * 3.1.0    2021-10-04 [1] CRAN (R 4.1.0)
 usethis     * 2.1.3    2021-10-27 [1] CRAN (R 4.1.0)
 withr         2.4.2    2021-04-18 [1] CRAN (R 4.1.0)
 xfun          0.27     2021-10-18 [1] CRAN (R 4.1.0)
 XML           3.99-0.8 2021-09-17 [1] CRAN (R 4.1.0)
 xml2          1.3.2    2020-04-23 [1] CRAN (R 4.1.0)

 [1] /Library/Frameworks/R.framework/Versions/4.1/Resources/library
`````
