### 2022-01-25 43df95a
Fix files not required to built. Update .Rbuildignore
