+++ R/CetesbRetrieveMet.R
@@ -13,7 +13,7 @@
 #' @param verbose Print query summary
 #' @param to_csv  Creates a csv file. FALSE by default
 #'
-#' @return data.frame wth Temperature (C), Relative Humidity (%),
+#' @return data.frame with Temperature (C), Relative Humidity (%),
 #' Wind Speed (m/s) and Direction (degrees), and Pressure information.
 #' @export
 #'
