+++ R/CetesbRetrievePol.R
@@ -12,7 +12,7 @@
 #' @param verbose Print query summary
 #' @param to_csv  Creates a csv file. FALSE by default
 #'
-#' @return data.frame wth O3, NO, NO2, PM2.5, PM10 and CO information.
+#' @return data.frame with O3, NO, NO2, PM2.5, PM10 and CO information.
 #' Units are ug/m3 except for CO which is in ppm, and NOx which is in ppb.
 #' @export
 #'
