+++ man/MonitorArRetrieveMetPol.Rd
@@ -33,7 +33,7 @@ O3, NO, NO2, NOx, PM2.5, PM10 and CO information.
 }
 \description{
 This function download the main meteorological parameters for model
-evaluation, together with criteria pollutatns for in air quality station
+evaluation, together with criteria pollutants for in air quality station
 (AQS) of Monitor Ar program.
 It will pad out the date with missing data with NA
 }
