### 2022-01-25 bcce5b5
Add Code of conduct in README
