### 2022-01-25 9e4d7f0
Add Zenodo DOI to README and CITATION and updating site
