### 2022-01-26 0c90cfd
Console message using message() instead of cat
