+++ R/WriteCSV.R
@@ -9,13 +9,19 @@
 #' @noRd
 #' @keywords internal
 
-WriteCSV <- function(aqs_data_df, aqs_name, start_date, end_date, vars){
+WriteCSV <- function(aqs_data_df, aqs_name, start_date, end_date,
+                     vars, csv_path){
   file_name <- paste0(aqs_name, "_",
                       paste0(vars, collapse = "_"), "_",
                       gsub("/", "-", start_date), "_",
                       gsub("/", "-", end_date), ".csv")
-  utils::write.table(aqs_data_df, file_name, sep = ",", row.names = FALSE)
 
-  file_path <- paste(getwd(), file_name, sep = "/")
+  if (csv_path == ""){
+    file_path <- paste(getwd(), file_name, sep = "/")
+  } else {
+    file_path <- paste(csv_path, file_name, sep = "/")
+  }
+
+  utils::write.table(aqs_data_df, file_path, sep = ",", row.names = FALSE)
   message(paste(file_path, "was created"))
 }
