+++ R/MonitorArRetrieveMetPol.R
@@ -11,6 +11,7 @@
 #' @param verbose Print query summary
 #' @param to_local Date information in local time. TRUE by default.
 #' @param to_csv Creates a csv file. FALSE by default.
+#' @param csv_path Path to save the csv file.
 #'
 #' @return data.frame with Temperature (C), Relative Humidity (%),
 #' Wind Speed (m/s) and Direction (degrees), Pressure information (hPa),
@@ -27,7 +28,7 @@
 #' }
 MonitorArRetrieveMetPol <- function(start_date, end_date, aqs_code,
                                     verbose = TRUE, to_local = TRUE,
-                                    to_csv = FALSE){
+                                    to_csv = FALSE, csv_path = ""){
   # Check is aqs_code is valid
   if (!(aqs_code %in% aqs_monitor_ar$code)){
     stop("Wrong aqs_code, please check monitor_ar_aqs", call. = FALSE) # nocov
@@ -55,7 +56,7 @@ MonitorArRetrieveMetPol <- function(start_date, end_date, aqs_code,
   names(all_data)[13] <- "pm25"
 
   if (to_csv){
-    WriteCSV(all_data, aqs_name, start_date, end_date, "MET_POL") # nocov
+    WriteCSV(all_data, aqs_name, start_date, end_date, "MET_POL", csv_path) # nocov
   }
 
   return(all_data)
