+++ R/MonitorArRetrieveMet.R
@@ -10,6 +10,7 @@
 #' @param verbose Print query summary.
 #' @param to_local Date information in local time. TRUE by default.
 #' @param to_csv Create a csv file. FALSE by default.
+#' @param csv_path Path to save the csv file.
 #'
 #' @return data.frame with Temperature (c), Relative Humidity (%),
 #' Wind speed (m/s) and direction (degrees) and Pressure information.
@@ -25,7 +26,7 @@
 #' }
 MonitorArRetrieveMet <- function(start_date, end_date, aqs_code,
                                  verbose = TRUE, to_local = TRUE,
-                                 to_csv = FALSE){
+                                 to_csv = FALSE, csv_path = ""){
   # Check is aqs_code is valid
   if (!(aqs_code %in% aqs_monitor_ar$code)){
     stop("Wrong aqs_code, please check monitor_ar_aqs", call. = FALSE) # nocov
@@ -50,7 +51,7 @@ MonitorArRetrieveMet <- function(start_date, end_date, aqs_code,
   names(all_met) <- c("date", "tc", "rh", "ws", "wd", "p", "aqs")
 
   if (to_csv){
-    WriteCSV(all_met, aqs_name, start_date, end_date, "MET") # nocov
+    WriteCSV(all_met, aqs_name, start_date, end_date, "MET", csv_path) # nocov
   }
 
   return(all_met)

[…]
