+++ man/CetesbRetrieveParam.Rd
@@ -12,7 +12,8 @@ CetesbRetrieveParam(
   start_date,
   end_date,
   verbose = TRUE,
-  to_csv = FALSE
+  to_csv = FALSE,
+  csv_path = ""
 )
 }
 \arguments{
@@ -32,6 +33,8 @@ to download}
 \item{verbose}{Print query summary}
 
 \item{to_csv}{Creates a csv file. FALSE by default}
+
+\item{csv_path}{Path to save the csv file}
 }
 \value{
 data.frame with parameters described in \code{params} vector
