+++ man/MonitorArRetrieveMet.Rd
@@ -10,7 +10,8 @@ MonitorArRetrieveMet(
   aqs_code,
   verbose = TRUE,
   to_local = TRUE,
-  to_csv = FALSE
+  to_csv = FALSE,
+  csv_path = ""
 )
 }
 \arguments{
@@ -25,6 +26,8 @@ MonitorArRetrieveMet(
 \item{to_local}{Date information in local time. TRUE by default.}
 
 \item{to_csv}{Create a csv file. FALSE by default.}
+
+\item{csv_path}{Path to save the csv file.}
 }
 \value{
 data.frame with Temperature (c), Relative Humidity (\%),
