### 2022-01-27 b0d7665
Fix test when using csv_path
