### 2022-01-29 85cc521
Update tests nrow, class, means
