+++ tests/testthat/test-CetesbRetrieveParam.R
@@ -9,7 +9,9 @@ test_that("CetesbRetrieveParam works!", {
                                    "01/01/2020", "07/01/2020",
                                    to_csv = TRUE, csv_path = csv_path)
 
-  expect_equal(TRUE, file.exists(paste0(
-    csv_path, "/Pinheiros_O3_NOX_VV_01-01-2020_07-01-2020.csv")
+  expect_equal(TRUE,
+               file.exists(
+                 paste0(csv_path,
+                        "/Pinheiros_O3_NOX_VV_01-01-2020_07-01-2020.csv")
     ))
 })
