### 2022-01-29 cb808be
Ensuring columns with data as numeric
