+++ tests/testthat/test-MonitorArRetrieveMetPol.R
@@ -10,7 +10,13 @@ test_that("multiplication works", {
   expect_equal(TRUE,  "numeric" %in% class(ca_all$rh))
   expect_equal(TRUE,  "numeric" %in% class(ca_all$ws))
   expect_equal(TRUE,  "numeric" %in% class(ca_all$wd))
+  expect_equal(TRUE,  "numeric" %in% class(ca_all$p))
   expect_equal(TRUE,  "numeric" %in% class(ca_all$o3))
+  expect_equal(TRUE,  "numeric" %in% class(ca_all$no))
+  expect_equal(TRUE,  "numeric" %in% class(ca_all$no2))
+  expect_equal(TRUE,  "numeric" %in% class(ca_all$nox))
   expect_equal(TRUE,  "numeric" %in% class(ca_all$co))
+  expect_equal(TRUE,  "numeric" %in% class(ca_all$pm10))
+  expect_equal(TRUE,  "numeric" %in% class(ca_all$pm25))
   expect_equal(TRUE,  "character" %in% class(ca_all$aqs))
 })
