+++ vignettes/qualr.Rmd
@@ -20,12 +20,12 @@ knitr::opts_chunk$set(
 # Context
 Both the State of São Paulo and Rio de Janeiro have one of the most extensive 
 air quality stations network in South America.
-[CETESB QUALAR System]() provide to the user the air quality data from the State
+[CETESB QUALAR System](https://cetesb.sp.gov.br/ar/qualar/) provide to the user the air quality data from the State
 of São Paulo. QUALAR System limits the download to one parameter from one air 
 station for one year in a simple query (three parameter in advance query).
 The data can have missing hours (e.g. due to calibration), 
 the decimal separator is ",", and the output is a CSV file.
-[data.rio]() hosts the air quality information from Monitor Ar Program.
+[data.rio](https://www.data.rio/maps/PCRJ::qualidade-do-ar-dados-hor%C3%A1rios/about) hosts the air quality information from Monitor Ar Program.
 It is not an user-friendly API and the data needs the same preprocessor as the 
 data from QUALAR System.
 
@@ -105,6 +105,7 @@ with `NA`) with a `date` column in `POSIXct`. This ensure compatibility with the
 Here is the code to  use openair `timeVariation()` function. Note that no 
 preprocessing is needed.
 ```{r eval=FALSE}
+library(openair)
 openair::timeVariation(rj_centro, pollutant = "PM10")
 ```
 
