### 2022-01-31 30b61db
Add qualR.R
