+++ R/qualR.R
@@ -1,0 +1,18 @@
+#' qualR: An R package to download Sao Paulo and Rio de Janeiro
+#' air pollution data
+#'
+#' @name qualR
+#' @author Mario Gavidia-Calderon and Daniel Schuch and Maria de Fatima Andrade
+#' @docType package
+#'
+#' @seealso
+#'\strong{Useful links:}
+#' \itemize{
+#' \item{Development repository:
+#'  \url{https://github.com/quishqa/qualR}}
+#' \item{Report bugs:
+#'  \url{https://github.com/quishqa/qualR/issues}}
+#' \item{qualR site:
+#'  \url{https://quishqa.github.io/qualR/}}
+#' }
+"_PACKAGE"
