### 2022-01-31 4e681b9
Add the account requirement in cetesb_retrieve functions
