+++ man/cetesb_retrieve_met_pol.Rd
@@ -42,6 +42,9 @@ This function download the main meteorological parameters for model
 evaluation, together with criteria pollutants for one
 air quality station (AQS). It will pad out the date with
 missing data with NA.
+This function required to have
+\href{https://seguranca.cetesb.sp.gov.br/Home/CadastrarUsuario}{an account}
+in \href{https://qualar.cetesb.sp.gov.br/qualar/home.do}{CETESB QUALAR}.
 }
 \examples{
 \dontrun{
