+++ man/cetesb_retrieve_param.Rd
@@ -40,9 +40,12 @@ to download}
 data.frame with parameters described in \code{params} vector
 }
 \description{
-This function downloads the parameters in a vector, this parameters
+This function downloads the parameters in a vector. These parameters
 can be both pollutants or meteorological observations for one air
 quality station (AQS). It will pad out the date with missing data
+with NA. This function requires to have
+\href{https://seguranca.cetesb.sp.gov.br/Home/CadastrarUsuario}{an account}
+in \href{https://qualar.cetesb.sp.gov.br/qualar/home.do}{CETESB QUALAR}.
 }
 \examples{
 \dontrun{
