+++ man/cetesb_retrieve_met.Rd
@@ -40,6 +40,9 @@ Wind Speed (m/s) and Direction (degrees), and Pressure information.
 This function download the main meteorological parameters for
 model evaluation from one air quality station (AQS) of CETESB AQS network.
 It will pad out the date with missing data with NA.
+This function requires to have
+\href{https://seguranca.cetesb.sp.gov.br/Home/CadastrarUsuario}{an account}
+in \href{https://qualar.cetesb.sp.gov.br/qualar/home.do}{CETESB QUALAR}.
 }
 \examples{
 \dontrun{
