### 2022-01-31 fbbe18a
Update openair example in vignettes
