### 2022-01-31 406cfc6
Add Qualar and Monitor Ar links
