+++ DESCRIPTION
@@ -17,8 +17,10 @@ Authors@R: c(
            role = c("aut","ctb"),
            email = "d.schuch@northeastern.edu",
            comment = c(ORCID = "0000-0001-5977-4519")))
-Description: A package to download information from CETESB QUALAR and
-    MonitorAr systems. It contains function to download different parameters, a set of
+Description: A package to download information from 
+    CETESB QUALAR <https://cetesb.sp.gov.br/ar/qualar/> and
+    MonitorAr <https://www.rio.rj.gov.br/web/smac/monitorar-rio1> 
+    systems. It contains function to download different parameters, a set of
     criteria pollutants and the most frequent meteorological parameters used in
     air quality data analysis and air quality model evaluation.
 License: MIT + file LICENSE
