### 2022-01-31 985714f
Fixing conditional in cetesb_retrive
