### 2022-02-01 4a78c93
Add ggplot2 example in vignette
