### 2022-02-01 5748004
Update document qualR.R
