### 2022-02-01 ffb16fc
Remove Travis CI badge
