### 2022-02-01 #474 @quishqa
Sorry for the (very) late reply. Thanks a lot to @ldecicco-USGS and to @beatrizmilz and @kauedesousa for the
very exhaustive review. Now the package is definitely better and we also learned a lot
about building packages in R.

Here are the answers to each observation following the commit where we fix them.

## Response to @beatrizmilz

- [X] Consistent naming styles: it would be better to choose a language (English), and a coding style (ex. snake case) and use them in all function names, arguments, and names of columns in the datasets returned.

**Response:** Now all function names are snake case and the names of columns are the most common in atmospheric sciences.
See https://github.com/quishqa/qualR/commit/10f2dde71f2be0d027ecf1b93df274ff2a06f5d1
And https://github.com/quishqa/qualR/commit/388c779d95b6dc5c35bb0ce118bda3aa1bfbea4d

- [x] The tests can be improved. Consider also using GitHub Actions to perform an automated check.

**Response:** We updated the tests to check also number of rows, to check if output is written, to check if variables (a pollutant) mean is inside a range, and to check the class of each columns.
See https://github.com/quishqa/qualR/commit/85cc521c9d80dd3ab1639d00dbbfd96fe43292a8
And
https://github.com/quishqa/qualR/commit/036a8e1a5041238618625cb4bac4eab07e87170e

- [x] Files in the root directory: some of the files do not seem fit to be stored there (such as the files related to the paper), and some need to be ignored in the build of the package. I wrote in more detail in the report.

**Response:** Paper folder moved to `inst/paper` https://github.com/quishqa/qualR/commit/2d03ff4bf2b604decf4073d3806c6dabc9be4a4f
Also we updated our `.Rbuildignore` file see: https://github.com/quishqa/qualR/commit/43df95adefdbbb90f16802ad6b7a4cf14bf65d70

- [x] The spell-check found some typos that could be corrected.

**Response:** Corrected in https://github.com/quishqa/qualR/commit/57d14d04b72164dbe7fcddb0c79148330812d247

- [x] I wrote in the report some ideas for the vignettes, but they are only suggestions to make it easier for people to use the data.

**Response:** We are not used to work with tidyverse. Nervertheless we add one example in the vignette. See https://github.com/quishqa/qualR/commit/4a78c9356b8e2dd7cec09844676db8d6abd377e2

- [x] Modifying directories using `setwd()` is not something recommended in the book R Packages

**Response:** By adding the csv_path argument we don't need to used `setwd()` anymore. See https://github.com/quishqa/qualR/commit/f7f80553c0fdc063d65cee7dece6fec68dd35c07
and https://github.com/quishqa/qualR/commit/b0d76654ae538727ae961a72c142acdaed75ef07

####  Other recommendations from full report

- [x] Add rOpenSci Review badge

**Response:** Added in https://github.com/quishqa/qualR/commit/e6af4d1b57e03db54737ea7dada1d99312c3ea40

- [x] The CITATION file exists. There are releases on GitHUb, but I did not find any reference to Zenodo. Reference. There is 0 results searching for qualR on Zenodo: https://zenodo.org/search?page=1&size=20&q=qualR . Read more how to do it and have a DOI for your package on the GitHub documentation.

**Response:** DOI created in https://github.com/quishqa/qualR/commit/9e4d7f03a245f50df6f9f7920eba493546e73b7c

- Code of conduct: in the book, is recommended to use a code of conduct (CoC), and I did not find that information in the README.

**Response:** Created in https://github.com/quishqa/qualR/commit/bcce5b5fc1f4c69a027db8b915c421cc86288c42

- [X] `R/sysdata.rda` should be used to internal datasets.

**Response:** Yes these dataset is for internal use. It has the name of stations with and without diacritics. Using Tatu**i** is the same as using Tatu**í**

- [x] Please review the links in the vignette, there is broken links (such as [CETESB QUALAR System](), [data.rio]()).

**Response:** Corrected in https://github.com/quishqa/qualR/commit/f3d4d2ab664453e3e0b89db15c60fc24dd0a6881

- [x] `MonitorArRetrievePol()`, `MonitorArRetrieveMet()` and `CetesbRetrieveMetPol()` - The names of the columns of the dataset returned are in a good pattern (all in English and snake case). The column classes I think it would be better if the numeric variables (such as the air parameters) returned numeric/double (now some of them are lgl because there is only NA’s).

**Response:** Columns names from returned dataset standarized. See  https://github.com/quishqa/qualR/commit/388c779d95b6dc5c35bb0ce118bda3aa1bfbea4d .
Now numeric variables are returned as `numeric` even when there is only `NA`. See https://github.com/quishqa/qualR/commit/cb808be11d78a44a15129520480b1b2eb344032d

- [X] Console messages - It is not recommended to use `cat()` to communicate with the user of the functions

**Response:** Now we use `message()`. See https://github.com/quishqa/qualR/commit/0c90cfdda41b5ac66b3a5c19f34a366574252c77

- [X] IDEA - When the function is storing the csv with `to_csv = TRUE`, it would be nice to have an argument to write the path where the csv will be saved.

**Response:** Accepted in https://github.com/quishqa/qualR/commit/f7f80553c0fdc063d65cee7dece6fec68dd35c07. It helped to avoid using `setwd()` on tests.

- [X] IDEA - I think it would be nice to be able to search more than one station in the functions.

**Response:** We put a way to do it in the README section in https://github.com/quishqa/qualR#a-variable-from-all-cetesb-aqs

- [x] All the tests passed. There are only a few tests. I recommend the authors to write more expectations with expect_*(), such as: expect the class of the result, the class of the columns in the dataset.

**Response:** Now we evaluate classes. See https://github.com/quishqa/qualR/commit/85cc521c9d80dd3ab1639d00dbbfd96fe43292a8

- [x] There is a test are not testing anything related to the package: In test-CetesbRetrieveParam.R, the only expectation is `expect_equal(2 * 2, 4)`.

**Response:** Corrected in https://github.com/quishqa/qualR/commit/85cc521c9d80dd3ab1639d00dbbfd96fe43292a8

- [x] There is no GitHub action to check the package consistently.

**Response:** Added in https://github.com/quishqa/qualR/commit/60592c9c12d82dd1d87e9323170f4e0b728442a7

## Response to @kauedesousa

- [x] Datasets `monitor_ar_aqs` and `monitor_ar_param` don't match with the general naming used for the functions. I suggest that you rename these files to match with your naming logic. You may rename like, `ArStation` and `ArParameters`

 **Response:** Now we changed all function names using snake case. See https://github.com/quishqa/qualR/commit/10f2dde71f2be0d027ecf1b93df274ff2a06f5d1

- [X] Same for the CETESB data sets

**Response:** Now returned datasets have the most common names in atmospheric sciences. See https://github.com/quishqa/qualR/commit/388c779d95b6dc5c35bb0ce118bda3aa1bfbea4d

- [x] Your vignette don't run smoothly because you call the package `openair` with `::` without explicitly mentioning it with `library(openair)`. You should call the package with `library()`.

**Response:** Added in https://github.com/quishqa/qualR/commit/f3d4d2ab664453e3e0b89db15c60fc24dd0a6881

- In line 114 of` /R/CetesbRetrieve.R` you start with an `if()` and then a else with another `if()`. I think that these statements are very difficult to understand and to fix. Maybe you can use only `if()s` and stress on the possible outcomes that the call may return to you.

**Response:** Fixed in https://github.com/quishqa/qualR/commit/985714ffb38e9ffd1147e982649e698669f0bec2

- [x] You test if the functions are returning the right dimensions of your expected data frame. But you don't test whether the outcome is the correct data. Maybe you can download a small data set from the server and put it in the `sysdata.rda` and validate it with the output of your functions.

**Response:** We updated the tests to check also number of rows, to check if output is written, to check if variables (a pollutant) mean is inside a range, and to check the class of each columns.
See https://github.com/quishqa/qualR/commit/85cc521c9d80dd3ab1639d00dbbfd96fe43292a8
And
https://github.com/quishqa/qualR/commit/036a8e1a5041238618625cb4bac4eab07e87170e

- [x] Fix the typo in line 16 of file `/R/CetesbRetrieveMet.R`

**Response:** Corrected in https://github.com/quishqa/qualR/commit/57d14d04b72164dbe7fcddb0c79148330812d247

- [x] I recommend to put the paper files in `/inst/paper` so your main directory is more organized. Same for the png files.

**Response:** Moved in https://github.com/quishqa/qualR/commit/2d03ff4bf2b604decf4073d3806c6dabc9be4a4f

- [x] Remove README.Rmd and keep only README.md

**Response:**  Removed in https://github.com/quishqa/qualR/commit/b32c6ffd0170311d4010ff731899c614092bf7e4

- [x] /data-raw/ could be placed inside /data/

**Response:** That folder was created when using the function `usethis::use_data_raw()` during the package development.

- [ ] Remove `LICENSE.md`

**Response:** We still have doubts about which one we should remove LICENSE.md has more information.

- [x] In the DESCRIPTION file, put the links for ETESB QUALAR and MonitorAr systems if any inside a <>, like this [https://cetesb.sp.gov.br/ar/qualar/](https://cetesb.sp.gov.br/ar/qualar/) is very likely that CRAN will ask you to do that.

**Response:** Added in https://github.com/quishqa/qualR/commit/406cfc639d74bea67f6b8a51c4a29f0b48462083

- [x] Make a file called qualR.R in the \R directory, it makes a nice description of your package.

**Response:** Created in https://github.com/quishqa/qualR/commit/30b61db4b926c5b0d86c2b249ce758bbfc43c8a2
