### 2022-02-01 #474 @beatrizmilz
Hi @quishqa ! I'm happy to know that the review was usefull. Congratulations for the extensive work!
Seems like the website was not fully updated. Some of the pages on the reference are returning error 404, in the case when the function names were changed: https://quishqa.github.io/qualR/reference/cetesb_retrieve_param.html
Probably using `pkgdown::build_site()` would fix it :)
