### 2022-02-01 6c032e6
Manually linkng function references
