### 2022-02-01 #474 @quishqa
Thanks for noticing this. It's weird because the manual from the datasets are OK  but only the functions (which the names were changed)  are presenting this error. I already do `pkgdown::build_site()` . Locally it works but on the deployed web it does not work. I'm going to do a little more research.
