+++ _pkgdown.yml
@@ -15,6 +15,7 @@ reference:
   - monitor_ar_retrieve_pol
   - monitor_ar_retrieve_met
   - monitor_ar_retrieve_met_pol
+  - CetesbRetrievePol
 - title: "About"
   contents:
   - qualR
