+++ NAMESPACE
@@ -1,5 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
+export(CetesbRetrievePol)
 export(cetesb_retrieve_met)
 export(cetesb_retrieve_met_pol)
 export(cetesb_retrieve_param)
