### 2022-02-04 #474 @quishqa
Dear all, the returning 404 error was fixed in https://github.com/quishqa/qualR/commit/6e4ce0d80cbf71272905b283ae43480c44a7cad2
