### 2022-02-05 #474 @beatrizmilz
Hi Mario! I checked here and the website is working :) Congratulations.

Congratulations for the extensive work.

About the vignette, I'm glad that you added an example with ggplot2! ggplot2 is an great package to create visualizations, and enables us to create elegant graphs with the data that qualR returns.

About the other suggestions and comments that I wrote, seems like they all have been considered. `devtools::check()` passed as well.

## Tests
I still think that the tests could be improved. I wrote some comments and I hope they can be helpfull if you want to expand the testing in the package:

- As @kauedesousa have said, is good to test whether the outcome is the correct data. One function that you can try is [expect_snapshot()](https://testthat.r-lib.org/articles/snapshotting.html): in the first time you use it, it will save the expected outcome and will check in the future if the results are still the expected, comparing with what was saved.

- It seems to me that not all functions have tests consistently: some of the functions have more extensive tests and others tests only the number of rows and columns for example. Having good tests in all functions helps to find bugs in the future.

- Also, I think is a good idea to improve the name of the tests, because if it breaks, is easier to understand which is in the check log (some of them are `test_that("multiplication works", {`). As said in R Packages book, "The main thing is that the message associated with the test should be informative so that you can quickly narrow down the source of the problem." [(reference)](https://r-pkgs.org/tests.html).  But this should be a quick fix.

- I really like [this chapter about tests](https://r-pkgs.org/tests.html) in the  R Packages book!

 ---
 Again, congratulations Mario!
