### 2022-02-15 4bcee7d
Update cetesb_retrieve_pol test
