### 2022-02-15 b5e5435
Fixing aqs as character
