### 2022-02-15 29eb915
Completing cetesb_retrieve function tests
