### 2022-02-15 734666b
Fix date as posixct not numeric
