### 2022-02-15 3d3aaf3
Update monitor_ar_retrieve functions tests
