+++ tests/testthat/test-monitor_ar_retrieve_met_pol.R
@@ -4,7 +4,22 @@ test_that("multiplication works", {
   end_date   <- "07/01/2020"
   ca_all     <- monitor_ar_retrieve_met_pol(start_date, end_date, "CA")
 
+  # Testing data.frame dims
   expect_equal(ncol(ca_all), 14)
+  expect_equal(TRUE, ncol(ca_all) <= 145)
+
+  # Testing data.frame values
+  all_means<- colMeans(ca_all[, 2:13], na.rm = TRUE)
+  expect_equal(all_means["tc"] > 28, all_means["tc"] < 29)
+  expect_equal(all_means["rh"] > 74, all_means["rh"] < 75)
+  expect_equal(all_means["ws"] > 0.8, all_means["ws"] < 0.9)
+  expect_equal(all_means["wd"] > 158, all_means["wd"] < 159)
+  expect_equal(all_means["p"] > 1007, all_means["p"] < 1008)
+  expect_equal(all_means["o3"] > 33, all_means["o3"] < 35)
+  expect_equal(all_means["co"] > 0.5, all_means["co"] < 0.6)
+  expect_equal(all_means["pm10"] > 21, all_means["pm10"] < 22)
+
+  # Testing data.frame classes
   expect_equal(TRUE,  "data.frame" %in% class(ca_all))
   expect_equal(TRUE,  "numeric" %in% class(ca_all$tc))
   expect_equal(TRUE,  "numeric" %in% class(ca_all$rh))
