### 2022-02-15 0bcfb60
Fixing nrow for monitor_ar_retrieve for windows
