### 2022-02-15 #474 @quishqa
Thanks again for the feedback!  We updated the test for all the functions we evaluated dims. values and classes,
see https://github.com/quishqa/qualR/commit/29eb9151387f86686f917f23261a7608a9c92d1c
and https://github.com/quishqa/qualR/commit/3d3aaf3755dba82e0a35bfc6411954017a660540

Also, we updated the test messages in https://github.com/quishqa/qualR/commit/0a6b1481afae9bbe802f92e1175e17e325d5529d
