### 2022-03-05 #474 @kauedesousa
No objection. I recommend the approval. I just wonder why you didn't submit it with a paper for JOSS. Anyways, thanks for opportunity to review the R package.
