### 2022-03-18 #474 @quishqa
Hi @maelle, actually we do. To do the first task we only need to follow [this](https://docs.github.com/en/repositories/creating-and-managing-repositories/transferring-a-repository), right? And we don't know where we need to do the `@ropensci-review-bot finalize transfer of <package-name>` comment?

Thanks!
