### 2022-03-18 #474 @maelle
Yes regarding the repo transfer!

Then the comment `@ropensci-review-bot finalize transfer of qualR` needs to be put here in an issue comment.

Thank you and have a good week-end!
