+++ README.md
@@ -378,7 +378,8 @@ That is, a concentration value for **22:00 is the mean from 21:01 to 22:00**.
 
 ## Code of Conduct
 
-Please note that the qualR project is released with a [Contributor Code of Conduct](https://contributor-covenant.org/version/2/0/CODE_OF_CONDUCT.html). By contributing to this project, you agree to abide by its terms.
+Please note that this package is released with a [Contributor Code of Conduct](https://ropensci.org/code-of-conduct/).
+By contributing to this project, you agree to abide by its terms.
 
 ## Acknowledgments
 
