+++ DESCRIPTION
@@ -32,7 +32,7 @@ Imports:
     XML,
     httr,
     jsonlite
-URL: https://github.com/ropensci/qualR
+URL: https://docs.ropensci.org/qualR (website) https://github.com/ropensci/qualR
 BugReports: https://github.com/ropensci/qualR/issues
 Suggests:
     knitr,
