### 2022-03-25 fabe98c
adding new installation method
