+++ tests/testthat/test-cetesb_retrieve_met_pol.R
@@ -22,7 +22,7 @@ test_that("cetesb_retrieve_met_pol works!", {
   expect_equal(param_means["nox"] > 20, param_means["nox"] < 21)
   expect_equal(param_means["co"] > 0.3, param_means["co"] < 0.4)
   expect_equal(param_means["pm10"] > 19, param_means["pm10"] < 20)
-  expect_equal(param_means["pm25"] > 5, param_means["pm25"] < 6)
+  expect_equal(param_means["pm25"] > 5, param_means["pm25"] < 7)
 
   # Testing data.frame classes
   expect_equal(TRUE,  "data.frame" %in% class(pinheiros))
